+++ src/components/paper/paper.html
@@ -1,0 +1,52 @@
+<g
+        class="paper"
+        :id="'paper' + id"
+        :transform="translateString"
+        style="display: block;"
+        @mouseover="$emit('bubbleMouseEnter', bubbleId)"
+        @mouseout="$emit('bubbleMouseLeave', bubbleId)"
+        @click="$emit('bubbleClick', bubbleId)"
+        @dblclick.prevent="$emit('bubbleDoubleClick', bubbleId)"
+>
+
+    <path
+            id="region"
+            :d="paperPath"
+            class="unframed"
+            style="fill-opacity: 1;">
+    </path>
+    <path
+            class="dogear"
+            :d="dogearPath"></path>
+
+    <foreignObject
+            id="article_metadata"
+            class="paperContent"
+            :width="w"
+            :height="h"
+            @mouseover="$emit('bubbleMouseEnter', bubbleId)"
+            @mouseout="$emit('bubbleMouseLeave', bubbleId)"
+            @click="$emit('bubbleClick', bubbleId)"
+            @dblclick.prevent="$emit('bubbleDoubleClick', bubbleId)"
+    >
+        <body>
+            <div class="paper_holder">
+
+                <div class="metadata" :style="{ height: metadata_height + 'px', width: metadata_width + 'px'}">
+                    <div id="icons">
+                        <p id="open-access-logo">&#xf09c;</p>
+                    </div>
+                    <p id="title" class="highlightable">
+                        {{ title }}<br />
+                    </p>
+                    <p id="details" class="highlightable">{{ authors_string }}</p>
+                    <p id="in">in <span class="highlightable">{{ published_in }} <span class="pubyear">({{ year }})</span></span></p>
+                </div>
+                <div class="readers">
+                    <p id="readers"><span id="num-readers">{{ readers }}</span><span class="readers_entity"> {{ base_unit }}</span></p>
+
+                </div>
+            </div>
+        </body>
+    </foreignObject>
+</g>
